+++ changelog.md
@@ -8,6 +8,7 @@
   * Attempting to use `Slice.downsampled_over()` or `Slice.downsampled_like()` with timestamps ranges or another channel that doesn't overlap with the channel now produces a `RuntimeError`.
   * Attempting to construct a `TimeSeries` where the length of the timestamp array is not equal to the length of the data array results in a `RuntimeError`.
   * `FdEnsemble` alignment now produces a `ValueError` if fewer than 2 datasets are provided.
+  * Image reconstruction now raises a `ValueError` if the length of the data and infowave are not equal.
 
 #### New features